Fix list counter overflow, add tests for dict encoding paths
Fix int overflow in encoder_listencode_list: the separator counter was
declared as `int` but should be `Py_ssize_t` to match the dict encoder
and correctly handle lists exceeding 2^31 items on 64-bit systems.

Add targeted test coverage for recent C extension changes:

- TestDictEncodingPaths: exercises the PyDict_Next fast path (unsorted
  exact dict), the iterator slow path (sorted, dict subclass), skipkeys
  on both paths, non-string keys, nested dicts, key_memo cache hits,
  and large dicts (1000 keys).

- Refcount leak tests: dict fast path, sorted slow path, and skipkeys
  on the fast path (INCREF/DECREF balance for borrowed PyDict_Next
  refs, encoder_encode_dict_key's Py_None sentinel, etc.).

+++ simplejson/_speedups.c
@@ -2922,7 +2922,7 @@ encoder_listencode_list(PyEncoderObject *s, JSON_Accu *rval, PyObject *seq, Py_s
     PyObject *iter = NULL;
     PyObject *obj = NULL;
     int is_true;
-    int i = 0;
+    Py_ssize_t i = 0;
 
     is_true = PyObject_IsTrue(seq);
     if (is_true == -1)
